fix data_generator error from having non-json files in data dirs
The code is sensitive to any file in the data dirs and wants to
treat it as json data.
Remove the git place holder .emptydir files for getting the
dirs in the template by way of git.
Create the dirs explicitly during the run as part of the
template setup.